+++ Products/CMFPlone/browser/admin.py
@@ -125,7 +125,6 @@ class AddPloneSite(BrowserView):
             'Products.CMFPlacefulWorkflow:CMFPlacefulWorkflow',
             'plone.app.registry:default',
             'plone.app.z3cform:default',
-            'plone.app.collection:default',
         ]
         utils = getAllUtilitiesRegisteredFor(INonInstallable)
         for util in utils: